refactor cli doctor check helpers
