+++ .sampo/changesets/issue-369-cli-doctor-check-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split doctor environment/template checks and workspace checks into focused runtime modules without changing doctor output.
